// web/src/api/generated/model/stockDocumentDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { StockDocumentKind } from './stockDocumentKind';
import type { StockLineDto } from './stockLineDto';

export interface StockDocumentDto {
  id: string;
  kind: StockDocumentKind;
  number: string;
  date: string;
  /** @nullable */
  memo: string | null;
  /** @nullable */
  counterAccountId: string | null;
  /** @nullable */
  voucherId: string | null;
  lines: StockLineDto[];
  value: number;
}